% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate.R
\name{animate}
\alias{animate}
\title{Animate a set of pngs}
\usage{
animate(
  dir,
  gif = "animation.gif",
  regexp = NULL,
  glob = NULL,
  delay = 2,
  loop = TRUE,
  max_size = 1200,
  repeat_last = 2
)
}
\description{
Animate a set of pngs
}
