% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bash.R
\name{bash}
\alias{bash}
\title{Launch git bash}
\usage{
bash()
}
\description{
This behaves the same way as launching git bash through the start menu,
explorer context menu or executable directly. The HOME environment variable
is adjusted, as R uses the Documents subfolder of the standard HOME directory
in git bash, so ".bash_profile" is correctly sourced on startup. The working
directory will correspond to that of the R session.
}
