% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{bro_duplicated}
\alias{bro_duplicated}
\title{Determine duplicate elements}
\usage{
bro_duplicated(x)
}
\arguments{
\item{x}{A vector of elements}
}
\description{
This function determines which elements of a vector or data frame exists
more than once, and returns a logical vector indicating those elements (rows) as duplicates.
In contrast to the base R function \code{duplicated()} it marks all occurences of the duplicated element,
including the first occurence.
}
