% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_plots.R
\name{bro_facet_wrap_paged}
\alias{bro_facet_wrap_paged}
\title{Dive a ggplot into facets to create a single or multi-page layout}
\usage{
bro_facet_wrap_paged(
  gg,
  facet_var,
  ncol = NULL,
  nrow = NULL,
  width = NULL,
  height = NULL,
  ...
)
}
\arguments{
\item{gg}{A ggplot or list of ggplots}

\item{facet_var}{Descr.}

\item{ncol}{Descr.}

\item{nrow}{Descr.}

\item{width}{Descr.}

\item{height}{Descr.}

\item{...}{Descr.}
}
\description{
Dive a ggplot into facets to create a single or multi-page layout
}
