% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_plots.R
\name{bro_get_ggsize}
\alias{bro_get_ggsize}
\title{Determine maximum plot dimensions}
\usage{
bro_get_ggsize(gg, units = c("mm", "cm", "in"))
}
\arguments{
\item{gg}{A ggplot or list of ggplots}
}
\value{
Returns a named vector with the maximum overall width and height found in the input.
}
\description{
Determine maximum plot dimensions
}
