% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_plots.R
\name{bro_ggsave}
\alias{bro_ggsave}
\title{Universal ggplot export function}
\usage{
bro_ggsave(gg = last_plot(), filename, device = NULL, path = NULL,
  scale = 1, width = NA, height = NA, units = c("in", "cm", "mm"),
  dpi = 300, limitsize = TRUE, ...)
}
\arguments{
\item{gg}{ggplot or list of ggplots}

\item{filename}{Filename for export}

\item{device}{Device to use. Can either be a device function (e.g. png()), or one of "eps", "ps", "tex" (pictex), "pdf", "jpeg", "tiff", "png", "bmp", "svg" or "wmf" (windows only).}

\item{path}{Path to save plot to (combined with filename).}

\item{scale}{Multiplicative scaling factor.}

\item{dpi}{Plot resolution. Also accepts a string input: "retina" (320), "print" (300), or "screen" (72). Applies only to raster output types.}

\item{limitsize}{When TRUE (the default), ggsave will not save images larger than 50x50 inches, to prevent the common error of specifying dimensions in pixels.}

\item{...}{Other arguments passed on to the graphics device function, as specified by device.}
}
\description{
This function takes a ggplot (for single page) or list of ggplots (for multi page) and writes them to file.
In case the input has absolute dimensions (e.g. as a result of egg:set_panel_size or
patchwork::plot_layout), width and height of the outpu are adjusted to fit the content.
}
