% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_plots.R
\name{bro_set_panel_size}
\alias{bro_set_panel_size}
\title{Set absolute panel size for ggplots}
\usage{
bro_set_panel_size(gg, width = unit(40, "mm"), height = unit(40, "mm"))
}
\arguments{
\item{gg}{A ggplot or list of ggplots}

\item{width}{Descr.}

\item{height}{Descr.}
}
\description{
Set absolute panel size for ggplots
}
