% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_plots.R
\name{bro_wrap_plots_paged}
\alias{bro_wrap_plots_paged}
\title{Wrap mutiple plots in a (multipage) layout}
\usage{
bro_wrap_plots_paged(plotlist, ncol = 1, nrow = 1, width = NULL,
  height = NULL, ...)
}
\arguments{
\item{plotlist}{A list of ggplots}

\item{ncol}{Number of columns in the layout. If more plots are provided than fit on one page, a list of patchworks is returned.}

\item{nrow}{Number of rows in the layout. If more plots are provided than fit on one page, a list of patchworks is returned.}

\item{width}{Widths of individual plots. For absolute dimensions use somehting like \code{unit(40, "mm")}.}

\item{height}{Heights of individual plots. For absolute dimensions use somehting like \code{unit(40, "mm")}.}

\item{...}{Other arguments passed on to \code{patchwork::wrap_plots()}}
}
\description{
Wrap mutiple plots in a (multipage) layout
}
