% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broekx.R
\name{broekx_download}
\alias{broekx_download}
\title{Download photos from ouders.broekx.be}
\usage{
broekx_download(url = NULL, to = NULL)
}
\arguments{
\item{url}{Email link to a page with images. If NULL (default), the URL is
taken from the clipboard.}

\item{to}{Directory to save the images. If NULL (default), the current
working directory is used.}
}
\description{
Download photos from ouders.broekx.be
}
