% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\docType{data}
\name{connectors}
\alias{connectors}
\title{List of functions that can be used to load and save different types
of data}
\format{
An object of class \code{list} of length 12.
}
\usage{
connectors
}
\description{
List of functions that can be used to load and save different types
of data
}
\author{
Guilherme Salome
}
\keyword{datasets}
