% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-gimme-album.R
\name{convert_gimme_album}
\alias{convert_gimme_album}
\title{Download high-resolution photos from gimme.eu based on a saved html}
\usage{
convert_gimme_album(path = getwd(), max_size = 4000)
}
\arguments{
\item{path}{folder where you saved the album html page, and the individual
photos should be put instead, defaults to current working
directory}

\item{max_size}{maximum number of pixels for image width and height, defaults
to 4000}
}
\description{
Download high-resolution photos from gimme.eu based on a saved html
}
\examples{
\dontrun{
convert_gimme_album("path/to/folder")
}
}
