% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create}
\alias{create}
\title{Create Project}
\usage{
create(
  name = getwd(),
  example = TRUE,
  renv = TRUE,
  git = TRUE,
  replace = FALSE
)
}
\arguments{
\item{name}{(string, optional) Name of the new project or relative/absolute path to a folder.
If unspecified, uses the current working directory as the project folder and name.}

\item{example}{(logical) If TRUE, copies example files to the project.}

\item{renv}{(logical) If TRUE, initializes a virtual environment using 'renv'.}

\item{git}{(logical) If TRUE, initializes a Git repository and adds an initial commit.}

\item{replace}{(logical) If FALSE and the specified folder already exists, stops and displays an error.
If TRUE, replaces an existing folder with the new project.}
}
\description{
Initializes a new project with a predefined folder structure and optional examples, virtual environment, and Git repository.
}
\details{
The function creates a new project directory, including the 'R' source code folder, 'inst' folder for parameters,
connections, and data catalog, 'data' folder with subdirectories for different data types, and a 'logs' folder.
It copies necessary files from the 'bro' package for configuration and data, and sets up a '.gitignore' file in
relevant data folders. If requested, it also copies example files ('costs.csv' and 'data.yaml'). Additionally,
it initializes a virtual environment using 'renv' and installs the 'bro' package from the specified GitHub repository.
Lastly, it initializes a Git repository and adds an initial commit.
}
