% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_execution_env.R
\name{create_execution_env}
\alias{create_execution_env}
\title{Create Execution Environment}
\usage{
create_execution_env(env = NULL)
}
\arguments{
\item{env}{An optional environment to use; if not provided, a new environment is created.}
}
\value{
Returns the initialized execution environment with loaded parameters, data registry, and data status.
}
\description{
Creates and initializes an execution environment, loading default parameters, data registry, and data status.
}
\details{
This function allows the creation of an execution environment, which is essential for managing parameters, data registry,
and data status in a data science project. If no environment is provided, a new environment is created. Default file paths
for parameters, data registry, and data status are specified and loaded into the environment. The loaded information includes
parameters from a YAML file, data registry from another YAML file, and data status using the bro package. Messages are
displayed indicating the successful loading of each component.
}
