% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_changes.R
\name{detect_changes}
\alias{detect_changes}
\title{Detect Changes in Node and Inputs}
\usage{
detect_changes(node, execution)
}
\arguments{
\item{node}{The node for which changes are to be detected.}

\item{execution}{The execution environment containing the status information for nodes and input data.}
}
\value{
Returns TRUE if changes are detected in the node or its inputs, FALSE otherwise.
}
\description{
Detects changes in a specified node and its inputs by comparing their current status with the previous status stored in the execution environment.
}
\details{
This function checks if the node definition has changed by calling the "detect_changes_node" function. If changes are detected in the node,
it returns TRUE. Otherwise, it checks each input of the node using the "detect_changes_input" function and determines if any input has changed.
If any input has changed, the function returns TRUE; otherwise, it returns FALSE.
}
