% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_changes_input.R
\name{detect_changes_input}
\alias{detect_changes_input}
\title{Detect Changes in Input}
\usage{
detect_changes_input(name, execution)
}
\arguments{
\item{name}{The name of the input data for which changes are to be detected.}

\item{execution}{The execution environment containing the status information for input data.}
}
\value{
Returns TRUE if changes are detected, FALSE otherwise.
}
\description{
Detects changes in a specified input data by comparing its current status with the previous status stored in the execution environment.
}
\details{
This function checks if the specified input data has undergone changes by comparing its current status with the previous status
stored in the execution environment. The update strategy specified in the data registry determines how the comparison is made.
If no previous status is available, it assumes that the data has been updated. If the update strategy is based on hashing, the
function calculates the hash of the data and compares it with the previous hash. If changes are detected, it returns TRUE; otherwise, FALSE.
}
