% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_changes_node.R
\name{detect_changes_node}
\alias{detect_changes_node}
\title{Detect Changes in Node}
\usage{
detect_changes_node(node, execution)
}
\arguments{
\item{node}{The node for which changes are to be detected.}

\item{execution}{The execution environment containing the status information for nodes.}
}
\value{
Returns TRUE if changes are detected, FALSE otherwise.
}
\description{
Detects changes in a specified node by comparing its current hash with the previous hash stored in the execution environment.
}
\details{
This function takes a node and the execution environment as inputs. It calculates the hash of the provided node using the
rlang::hash function. It then compares this new hash with the previous hash stored in the execution environment for the same node.
If the hashes differ, indicating changes, the function returns TRUE and displays a message about the detected changes.
Otherwise, it returns FALSE, indicating no changes.
}
