% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_imported_packages}
\alias{.get_imported_packages}
\title{Get imported packages from DESCRIPTION}
\usage{
.get_imported_packages()
}
\value{
Character vector of imported package names
}
\description{
Extracts package names from the Imports field of the
DESCRIPTION file
}
\keyword{internal}
