% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gallery.R
\name{gallery_items}
\alias{gallery_items}
\title{Write a hugo hugrid items.toml file based on a google drive folder}
\usage{
gallery_items(drive_folder, path = "data/items.toml")
}
\arguments{
\item{drive_folder}{full path to google drive folder without trailing slash}

\item{path}{path to the \code{items.toml} file to be created}
}
\description{
This function allows you to create an image gallery type of website, using
the \href{https://themes.gohugo.io/hugrid/}{hugo hugrid theme}, with minimal setup
in a google drive folder. Files in the folder starting with the
folder/project name are omitted. The rest is processed in order, and added to
the items.toml file. Everything works by linking to google drive, so files
are hosted there, but all files should hence be available through a shared
link. Details for setup are provided below.
}
\details{
Every \strong{image file} should start with a code, then a name, and finally a
type, separated with underscores (\emph{e.g.} \code{code_name_type.jpg}). Types can be
\code{thumb}, for the thumbnail images on the main page grid, \code{image} for the
image used for display when the thumbnail is clicked, and \code{url} for the file
that is linked to when clicking the button that is foreseen in the theme.

If other types are included, these are added as separate buttons, by
appending an html \if{html}{\out{<a>}}\if{html}{\out{</a>}} tag to the description (explained below). For the
button text, the type is  used, but underscores are replaced by spaces.

A \strong{csv file} should be included as well, with a column \code{code} that links
the rows with the image files, and columns \code{title} and \code{description}, which
are displayed next to the \code{image} type file.
}
