% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_context.R
\name{get_context}
\alias{get_context}
\title{Returns an environment with hooks, data catalog and parameters.}
\usage{
get_context(parameters = "parameters.yml", global = FALSE, verbose = TRUE)
}
\arguments{
\item{parameters}{(string, optional) Name of a parameters YML file stored in
the conf/base folder. File must begin with "parameters" and end with ".yml".
If unspecified, the default parameters file named "parameters.yml" is loaded
in the environment. If specified, loads the specified parameters file into
the environment instead.}

\item{global}{(boolean, optional) If TRUE, adds parameters, catalog and
hooks to the global environment.}

\item{verbose}{(boolean, optional) If TRUE, prints messages.}
}
\value{
(environment) Environment with hooks, catalog and parameters.
}
\description{
Creates a 'context' environment with hooks, data catalog and parameters.
This environment is used to separate the execution of the pipelines from
variables in the global environment. This environment can also be used to
debug your code, since it exposes the data catalog and can be used with the
other exported functions in the framebar.
}
\details{
The environment 'context' is created by loading a parameters file, either
conf/base/parameters.yml or another one specified by the user.

It also adds the following extra values to the list of parameters:
- 'current_date_and_time': the current date and time in the format
  "%Y%m%d_%H%M%S".
- 'parameters_file_name': the name of the parameters file used to create the
  context.
If either of these parameters are already present in the parameters file, an
error is thrown.

Then, it loads the data catalog file in conf/base/catalog.yml.
Then, it loads the credentials file in conf/local/credentials.yml if it
exists.
Finally, it sources the hooks file in src/project_name/hooks.R. When hooks
are sourced, all of the files containing pipelines, nodes and functions will
run, thus making all vailable in the environment. If multiple functions have
the same name, make sure that, in the hooks file, they are sourced to
different environments.
}
