% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get Data}
\usage{
get_data(name, execution)
}
\arguments{
\item{name}{The name of the dataset to be retrieved or loaded.}

\item{execution}{The execution environment containing the data registry and loaded datasets.}
}
\value{
Returns the specified dataset, either retrieved from the environment or loaded from the data registry.
}
\description{
Retrieves and returns a specified dataset from the execution environment or loads it if not already loaded.
}
\details{
This function checks if the requested dataset is a parameter; if so, it directly retrieves it from the execution environment.
If the dataset is not a parameter and is already loaded in the execution environment, it is retrieved from there.
If the dataset is neither a parameter nor loaded, the function uses the "load_data" function to load it from the data registry.
}
