% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load Data}
\usage{
load_data(name, execution)
}
\arguments{
\item{name}{The name of the dataset to be loaded, as specified in the data registry.}

\item{execution}{The execution environment containing the data registry.}
}
\value{
Returns the loaded dataset, and it is also stored in the execution environment for further use.
}
\description{
Loads and returns a specified dataset from the data registry in the given execution environment.
}
\details{
This function retrieves the data registry from the provided execution environment. It checks if the specified dataset
exists in the registry and if the path to the data is available. If the dataset or path is missing, an error is raised.
The function then loads the dataset using the specified data loader based on the data type defined in the registry.
Messages are displayed indicating the successful loading of the dataset along with its type and loading method.
}
