% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_nodes.R
\name{load_nodes}
\alias{load_nodes}
\title{Load Nodes from Configuration File}
\usage{
load_nodes(path = file.path("R", "nodes.R"), execution = new.env())
}
\arguments{
\item{path}{(string) The path to the configuration file containing nodes and other objects.}

\item{execution}{(environment) An optional execution environment to use; if not provided, a new environment is created.}
}
\value{
Returns a list of nodes loaded from the specified configuration file.
}
\description{
Loads nodes from a specified configuration file into a separate environment.
}
\details{
This function takes the path to a configuration file (default is "R/nodes.R") and creates a new environment (`objects`).
It then sources the specified file, executing its contents within the new environment. The function selects only objects
with the class 'node' from the environment, which are typically used to represent nodes in a data processing or analysis
pipeline. These nodes can include various data processing functions or other objects relevant to the workflow.
}
