% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{node}
\alias{node}
\title{Create a Node in a Data Processing Pipeline}
\usage{
node(f, x = NULL, y = NULL, name = deparse(substitute(f)))
}
\arguments{
\item{f}{(function) The R function that the node represents in the pipeline.}

\item{x}{(character, optional) Inputs for the node, represented as character vectors.}

\item{y}{(character, optional) Outputs for the node, represented as character vectors.}

\item{name}{(character, optional) The name of the node. If not provided, it defaults to the name of the function.}
}
\value{
Returns a node, which is a named list with components representing the function, inputs, outputs, and name.
}
\description{
Creates a node that represents a function and its inputs and outputs within a data processing or analysis pipeline.
}
\details{
This function allows the user to define a node in a pipeline, where a node consists of a function (`f`) along with
its inputs (`x`) and outputs (`y`). The function can be any R function, and the inputs and outputs are optional.
The name of the node can be specified using the `name` parameter, and if not provided, it defaults to the name of the function.
The resulting node is a named list with components `f` (function), `x` (inputs), `y` (outputs), and `name` (node name).
Nodes are typically used to represent individual steps or processes in a data workflow.
}
