% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{plan_week}
\alias{plan_week}
\title{Plain text week planning}
\usage{
plan_week(file = NULL, previous = NULL)
}
\arguments{
\item{file}{File to write to, or "clipboard".}
}
\value{
The text, invisibly if written first to a file.
}
\description{
Plain text week planning
}
