% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{plan_year}
\alias{plan_year}
\title{Plain text year planning}
\usage{
plan_year(file = NULL, year = lubridate::year(lubridate::today()))
}
\arguments{
\item{file}{File to write to, or "clipboard".}

\item{year}{The year to use, defaults to the current year.}
}
\value{
The text, invisibly if written first to a file.
}
\description{
Plain text year planning
}
