% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_variable.R
\name{retrieve_variable}
\alias{retrieve_variable}
\title{Retrieves a variable with a given name from memory or disk.}
\usage{
retrieve_variable(
  varname,
  parameters = "parameters.yml",
  context = NULL,
  force_reload = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{varname}{(string) Name of variable or data piece to retrieve.}

\item{parameters}{(string, optional) Name of a parameters YML file stored in
the conf/base folder. File must begin with "parameters" and end with ".yml".
If unspecified, the default parameters file named "parameters.yml" is loaded
in the environment. If specified, loads the specified parameters file into
the environment instead.}

\item{context}{(environment, optional) An environment containing a data
catalog and parameters. If unspecified, will attempt to create such an
environment from the current working directory (requires the current working
directory to follow framebar's file structure).}

\item{force_reload}{(boolean, optional) If TRUE, forces a reload of the
variable from disk, even if it is already in memory. If FALSE, the function
will check if the variable is already in memory and return it without
reloading it from disk. Default is FALSE.}

\item{verbose}{(boolean, optional) If TRUE, prints a message when
successfully loaded data.}
}
\value{
(object) Object containing data.
}
\description{
If the variable is not in memory, utilizes the catalog in context to find
the data on disk. If the data is not available, stops execution. If the
variable is in memory, this function can retrieve the variable itself, or
any of its elements (using the `$` operator).
}
