% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Executes all nodes listed in a pipeline.}
\usage{
run(
  hook = "default",
  parameters = "parameters.yml",
  cached = FALSE,
  overwrite = c(),
  rerun_intermediary = FALSE,
  clear = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{hook}{(string, optional) String used to recover the pipeline(s) from
the hooks list. Defaults to the 'default' pipeline.}

\item{parameters}{(string, optional) Name of a parameters YML file stored in
the conf/base folder. File must begin with "parameters" and end with ".yml".
If unspecified, uses the default parameters file named "parameters.yml". If
specified, all references to "parameters" within a node will use the
specified parameters file. Any output file that is saved will be stored
under a folder with same name as the parameters file. For example, an output
named "results.rds" generated with a parameters file named
"parameters_2020.yml" would be saved in as "parameters_2020.yml/results.rds".}

\item{cached}{(boolean, optional) If FALSE, executes each node. If TRUE,
only execute a node if its outputs do not exist in memory or in disk (local
or remote). Does not support versioned data.}

\item{overwrite}{(list, optional) List of outputs that will be overwritten
even in cached mode (i.e. the corresponding node will be executed even if
the output exists).}

\item{rerun_intermediary}{(boolean, optional) If TRUE, executes intermediary
nodes even when their outputs are not required in future nodes with cached
outputs.}

\item{clear}{(boolean, optional) If TRUE, clears variables from memory as
they stop being used by nodes. Always keep outputs from last node.}

\item{verbose}{(boolean, optional) If TRUE, prints information about each
node during execution.}
}
\value{
(environment) Environment containing the catalog, parameters,
connection and outputs from nodes.
}
\description{
Pipelines are a sequence of nodes that can be run in sequence.
Pipelines are given names in the hooks.R file.
Given a name of a hooked pipeline, framebar will execute each of its nodes
in a separate environment. The environment is returned at the end of the
execution. Each call to run creates and runs on a new environment.
}
