% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_node.R
\name{run_node}
\alias{run_node}
\title{Run Node}
\usage{
run_node(node, execution)
}
\arguments{
\item{node}{The node to be executed in the workflow.}

\item{execution}{The execution environment containing input data, output data, and status information.}
}
\value{
Returns the output of the executed node, stored in the execution environment.
}
\description{
Executes a specified node in the workflow, loading input data, running the node's function, and storing the output.
}
\details{
This function executes a specified node in the workflow. It first loads input data for the node using the "get_data" function.
It then executes the node's function using the loaded inputs and stores the output. If the node produces a single output,
the output is saved to a file and in memory using the "save_data" function. If the node produces multiple outputs, each output
is saved to a file and in memory separately. The status information is updated using the "update_status" function.
Messages are displayed indicating the start and end of node execution, along with the runtime.
}
