% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_dotfiles}
\alias{setup_dotfiles}
\title{Setup dotfiles}
\usage{
setup_dotfiles(scope = "user")
}
\arguments{
\item{scope}{Edit globally for the current \strong{user}, or locally for the
current \strong{project}}
}
\value{
Invisible NULL.
}
\description{
Setup dotfiles
}
