% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_r_profile}
\alias{setup_r_profile}
\title{Setup a .Rprofile file}
\usage{
setup_r_profile(scope = "user")
}
\arguments{
\item{scope}{Edit globally for the current \strong{user}, or locally for the
current \strong{project}}
}
\value{

}
\description{
Setup a .Rprofile file
}
