% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{shortcuts}
\alias{shortcuts}
\alias{pkg}
\alias{last}
\alias{serve}
\alias{git}
\title{Shortcuts for interactive use}
\usage{
pkg(name)

last()

serve()

git()
}
\arguments{
\item{name}{Package name to check}
}
\description{
Shortcuts for interactive use
}
