% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio.R
\name{toggle}
\alias{toggle}
\alias{light}
\alias{dark}
\title{Toggle between light and dark RStudio solarized themes}
\usage{
toggle()

light()

dark()
}
\description{
The {rsthemes} solarized themes are used here, as they provide more
consistent use of italics, better highlighting, etc. as the built-in
solarized themes.
}
