% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traverse_nodes.R
\name{traverse_nodes}
\alias{traverse_nodes}
\title{Traverse Nodes and Create Dependency Matrix}
\usage{
traverse_nodes(nodes)
}
\arguments{
\item{nodes}{(list) A list of nodes, where each node is represented by a named list with 'x' (inputs) and 'y' (outputs) components.}
}
\value{
Returns a dependency matrix indicating relationships between nodes, where 1 indicates dependency and
0 indicates no dependency. Each row indicates a different node, and the dependency relationship to the nodes in the columns.
}
\description{
Creates a dependency matrix based on the relationships between nodes provided.
}
\details{
Given a list of nodes, each represented by a named list with 'x' (inputs) and 'y' (outputs) components,
this function generates a dependency matrix. The matrix indicates dependencies between nodes, where a value of
1 indicates that a node depends on another node. Dependencies are determined by checking if any of the inputs
of a node are outputs of another node. The resulting matrix provides insights into the order in which nodes
should be executed in a data processing or analysis pipeline.
}
