% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_status.R
\name{update_status}
\alias{update_status}
\title{Update Status Information}
\usage{
update_status(node, inputs)
}
\arguments{
\item{node}{The node for which the status information is updated.}

\item{inputs}{The inputs used by the node, including data for which hash values need to be updated.}
}
\value{
Returns NULL after updating the status information.
}
\description{
Updates the status information with hash values for a specified node and its inputs.
}
\details{
This function updates the status information with hash values for a specified node and its inputs. It first loads the current
status information from the YAML file, then updates the hash value for the specified node and each input data used by the node.
Finally, it writes the updated status information back to the YAML file.
}
