% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_exists.R
\name{var_exists}
\alias{var_exists}
\title{Checks if a variable exists in memory or disk.}
\usage{
var_exists(output, context, parameters = "parameters.yml")
}
\arguments{
\item{output}{(string) Name of a variable.}

\item{context}{(environment) Environment containing data catalog.}

\item{parameters}{(string, optional) Name of a parameters YML file stored in
the conf/base folder.}
}
\value{
(boolean) TRUE if variable exists in memory or disk.
}
\description{
Given the name of a variable, check if it exists in memory or in disk (local
or remote).
}
