% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_status.R
\name{write_status}
\alias{write_status}
\title{Write Status to File}
\usage{
write_status(status, path = file.path("data", "status.yaml"))
}
\arguments{
\item{status}{The status information to be written to the YAML file.}

\item{path}{The file path where the status information will be stored in YAML format.}
}
\description{
Writes the provided status information to a YAML file at the specified path.
}
\details{
This function takes the provided status information and writes it to a YAML file at the specified path.
The status information typically includes the hash values for nodes and data, indicating the current state of the execution environment.
}
