% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_plots.R
\name{bro_ggsave}
\alias{bro_ggsave}
\title{Universal ggplot export function}
\usage{
bro_ggsave(gg = last_plot(), filename, device = NULL, path = NULL,
  scale = 1, width = NA, height = NA, units = c("in", "cm", "mm"),
  dpi = 300, limitsize = TRUE, ...)
}
\arguments{
\item{gg}{ggplot or list of ggplots}

\item{filename}{filename for export}

\item{width}{width of the pdf (default 7 inch). When the plot has absolute dimensions this parameter gets overridden to fit the content.}

\item{height}{height of the pdf (default 7 inch). When the plot has absolute dimensions this parameter gets overridden to fit the content.}

\item{...}{gets passed to pdf()}
}
\description{
This function takes a ggplot (for single page) or list of ggplots (for multi page) and writes them to a pdf file.
In case the input has absolute dimensions (e.g. as a result of egg:set_panel_size or
patchwork::plot_layout), width and height of the pdf are overridden to fit the content.
}
