% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_nodes.R
\name{load_nodes}
\alias{load_nodes}
\title{Load Nodes from a Script File}
\usage{
load_nodes(path = file.path("R", "nodes.R"), execution = new.env())
}
\arguments{
\item{path}{The file path to the script containing nodes and other objects (default: "R/nodes.R").}

\item{execution}{An optional execution environment to use; if not provided, a new environment is created.}
}
\value{
Returns the objects with the 'node' class loaded from the specified script file.
}
\description{
Loads nodes and other objects, including functions, from a specified script file into a separate environment.
}
\details{
This function takes a file path as input, which defaults to "R/nodes.R". It creates a new environment and sources the
script file within this environment. All objects, including functions, are stored in this separate environment.
The function then retrieves and returns only the objects with the 'node' class from the created environment.
These nodes are extracted from a dunder variable named "__nodes__", that is created when the "nodes.R" file
is sourced.
}
