% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystem.R
\name{path_fix}
\alias{path_fix}
\title{Fix backslashes in a windows file path}
\usage{
path_fix(path)
}
\arguments{
\item{path}{Path to process. If missing, the path is read from the clipboard.}
}
\value{
The fixed path, invisibly.
}
\description{
<kbd>Shift</kbd> + Right-click in the Windows file Explorer allows you to copy a file
path, but it has backslashes. With this function, you get the forward
slash version, so it can be used inside your R code.
}
